Update and rename adolf_ETIP.js to adolf_pk.js
